lib/acpi: Move the RSDP and table lookup to stage 2
